Fix upward scrolling bug when using scroll wheel
After selecting a Stack Overflow link, if the scroll bar is at the
top of the page, scrolling upwards using the scroll wheel causes the
scroll wheel to move to the bottom of the page instead of staying at
the top of the page.

This was fixed by adding a check to the scroll wheel up mouse event
and only doing ow.set_scrollpos(pos - 1) if pos > 0.