Add error for Stack Overflow get request
Currently there is no error handling for the Stack Overflow get
request in the souper function. If the requests.get function throws
an exception rebound just crashes.

Changed this by putting the requests.get call in a try except block
and adding a friendly error message if the requests.get call throws
an exception.